# backend/FeeCatalog/openapi/spec.yaml
openapi: 3.0.3
info:
  title: Fee Catalog API
  description: API for managing fee structures, add-ons, courses, categories, and
    batches for medical college admissions.
  version: 0.1.0
...
paths:
...
  /categories/{id}:
    get:
      tags:
        - categories
      description: Returns a single category by id.
      parameters:
        - schema:
            pattern: ^[1-9][0-9]*$
            type: string
          in: path
          name: id
          required: true
          description: Category id
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: object
                    required:
                      - type
                      - id
                      - attributes
                    properties:
                      type:
                        type: string
                        enum:
                          - categories
                      id:
                        type: string
                      attributes:
                        type: object
                        required:
                          - name
                        properties:
                          name:
                            type: string
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - errors
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      required:
                        - status
                        - code
                        - title
                        - detail
                      properties:
                        status:
                          description: HTTP status code, as a string
                          type: string
                        code:
                          description: Machine-readable domain error code
                          type: string
                        title:
                          description: Short, human-readable error name
                          type: string
                        detail:
                          description: Human-readable explanation specific to this error
                          type: string
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - errors
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      required:
                        - status
                        - code
                        - title
                        - detail
                      properties:
                        status:
                          description: HTTP status code, as a string
                          type: string
                        code:
                          description: Machine-readable domain error code
                          type: string
                        title:
                          description: Short, human-readable error name
                          type: string
                        detail:
                          description: Human-readable explanation specific to this error
                          type: string
...
  /courses/{id}:
    get:
      tags:
        - courses
      description: Returns a single course by id.
      parameters:
        - schema:
            pattern: ^[1-9][0-9]*$
            type: string
          in: path
          name: id
          required: true
          description: Course id
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: object
                    required:
                      - type
                      - id
                      - attributes
                    properties:
                      type:
                        type: string
                        enum:
                          - courses
                      id:
                        type: string
                      attributes:
                        type: object
                        required:
                          - name
                          - durationYears
                        properties:
                          name:
                            type: string
                          durationYears:
                            type: number
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - errors
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      required:
                        - status
                        - code
                        - title
                        - detail
                      properties:
                        status:
                          description: HTTP status code, as a string
                          type: string
                        code:
                          description: Machine-readable domain error code
                          type: string
                        title:
                          description: Short, human-readable error name
                          type: string
                        detail:
                          description: Human-readable explanation specific to this error
                          type: string
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - errors
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      required:
                        - status
                        - code
                        - title
                        - detail
                      properties:
                        status:
                          description: HTTP status code, as a string
                          type: string
                        code:
                          description: Machine-readable domain error code
                          type: string
                        title:
                          description: Short, human-readable error name
                          type: string
                        detail:
                          description: Human-readable explanation specific to this error
                          type: string
...
  /batches/{id}:
    get:
      tags:
        - batches
      description: Returns a single batch by id.
      parameters:
        - schema:
            pattern: ^[1-9][0-9]*$
            type: string
          in: path
          name: id
          required: true
          description: Batch id
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: object
                    required:
                      - type
                      - id
                      - attributes
                    properties:
                      type:
                        type: string
                        enum:
                          - batches
                      id:
                        type: string
                      attributes:
                        type: object
                        required:
                          - name
                          - startDate
                          - endDate
                        properties:
                          name:
                            type: string
                          startDate:
                            description: ISO 8601 date (YYYY-MM-DD)
                            type: string
                          endDate:
                            description: ISO 8601 date (YYYY-MM-DD)
                            type: string
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - errors
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      required:
                        - status
                        - code
                        - title
                        - detail
                      properties:
                        status:
                          description: HTTP status code, as a string
                          type: string
                        code:
                          description: Machine-readable domain error code
                          type: string
                        title:
                          description: Short, human-readable error name
                          type: string
                        detail:
                          description: Human-readable explanation specific to this error
                          type: string
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - errors
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      required:
                        - status
                        - code
                        - title
                        - detail
                      properties:
                        status:
                          description: HTTP status code, as a string
                          type: string
                        code:
                          description: Machine-readable domain error code
                          type: string
                        title:
                          description: Short, human-readable error name
                          type: string
                        detail:
                          description: Human-readable explanation specific to this error
                          type: string
...